requirements.txt
overload~=1.1
requests~=2.23.0
inputs~=0.5
opensimplex~=0.3
PyYAML~=5.3.1
fastutils~=0.10.1
pyglet~=1.5.5
Pillow~=7.1.2
future~=0.18.2
wxPython~=4.1.0
pyGoogleSearch~=2.5
Deprecated~=1.2.10
setuptools~=46.4.0
bizerror
dill~=0.3.1.1
git+https://github.com/ultrafunkamsterdam/googletranslate
